fix: Handle responsive for emails templates, reuse box and img as cpt
