+++ emails/community-new-event.tmpl.tsx
@@ -6,13 +6,14 @@ import {
   Head,
   Heading,
   Html,
-  Img,
   Preview,
   Row,
   Section,
   Text,
 } from "@react-email/components";
 import React from "react";
+import { EmailEventImg } from "./email-event-img";
+import { EmailEventBox } from "./email-event-box";
 
 // To generate an example, run `make generate && go run ./backend mail > event-announcement.html`
 
@@ -24,12 +25,7 @@ export const EventAnnouncementEmail = () => (
         New Event Sponsored by the Community - {"{{.EventName}}"}
       </Preview>
       <Container style={container}>
-        <Img
-          alt="Event presentation"
-          src="{{.EventImage}}"
-          width={960}
-          height={540}
-        />
+        <EmailEventImg src="{{.EventImage}}" />
         <Section style={welcome.section}>
           <Text style={welcome.text}>
             Join us for the upcoming event sponsored by {"{{.CommunityName}}"}!
@@ -47,16 +43,18 @@ export const EventAnnouncementEmail = () => (
             </Column>
           </Row>
           <Row>
-            <Column style={details.boxContainerLeft}>
-              <Box
+            <Column>
+              <EmailEventBox
                 title="DATE AND TIME"
                 icon={`{{.CalendarIconURL}}`}
                 iconAlt="Calendar icon"
                 content={`From: {{.EventDate}} To: {{.EventEndDate}}`}
               />
             </Column>
-            <Column style={details.boxContainerRight}>
-              <Box
+          </Row>
+          <Row>
+            <Column>
+              <EmailEventBox
                 title="SPONSORED BY"
                 icon={`{{.CommunityImage}}`}
                 iconAlt="Community logo"
@@ -77,40 +75,6 @@ export const EventAnnouncementEmail = () => (
   </Html>
 );
 
-const Box = (props: {
-  title: string;
-  icon: string;
-  iconAlt: string;
-  content: string;
-}) => {
-  return (
-    <Section style={box}>
-      <Row style={{ height: "100%" }}>
-        <Column style={{ height: "100%", verticalAlign: "top" }}>
-          <Section>
-            <Row>
-              <Column>
-                <Text style={boxTitle}>{props.title}</Text>
-              </Column>
-            </Row>
-            <Row>
-              <Column>
-                <Img
-                  alt={props.iconAlt}
-                  src={props.icon}
-                  width={32}
-                  height={32}
-                />
-                <Text style={boxContent}>{props.content}</Text>
-              </Column>
-            </Row>
-          </Section>
-        </Column>
-      </Row>
-    </Section>
-  );
-};
-
 export default EventAnnouncementEmail;
 
 const main = {
@@ -122,8 +86,7 @@ const main = {
 
 const container = {
   margin: "10px auto",
-  width: "600px",
-  maxWidth: "100%",
+  maxWidth: 800,
   border: "1px solid #F5F5F5",
 };
 
@@ -132,6 +95,7 @@ const welcome = {
     padding: "48px 20px",
     height: 220,
     backgroundColor: "#000000",
+    wordBreak: "break-word",
   },
   text: {
     color: "#FFFFFF",
@@ -142,38 +106,6 @@ const welcome = {
     lineHeight: 1.1,
     letterSpacing: -1.2,
   },
-} as const;
-
-const box = {
-  backgroundColor: "#F5F5F5",
-  borderRadius: 4,
-  padding: 12,
-  height: "100%",
-} as const;
-
-const boxTitle = {
-  margin: 0,
-  color: "#666666",
-  fontSize: 12,
-  fontWeight: 500,
-  letterSpacing: 0.5,
-  paddingBottom: 40,
-  lineHeight: 1.3,
-} as const;
-
-const boxContent = {
-  margin: 0,
-  fontSize: 16,
-  fontWeight: 500,
-  lineHeight: 1.3,
-  letterSpacing: -0.2,
-  paddingTop: 10,
-} as const;
-
-const boxContainerCommon = {
-  height: "100%",
-  width: "50%",
-  verticalAlign: "top",
 } as const;
 
 const details = {
@@ -195,14 +127,6 @@ const details = {
     letterSpacing: -0.6,
     margin: 0,
     marginBottom: 20,
-  },
-  boxContainerLeft: {
-    paddingRight: 8,
-    ...boxContainerCommon,
-  },
-  boxContainerRight: {
-    paddingLeft: 8,
-    ...boxContainerCommon,
   },
   seeEventButton: {
     backgroundColor: "#000000",
