+++ emails/event-cancelled.tmpl.tsx
@@ -5,13 +5,14 @@ import {
   Head,
   Heading,
   Html,
-  Img,
   Preview,
   Row,
   Section,
   Text,
 } from "@react-email/components";
 import React from "react";
+import { EmailEventImg } from "./email-event-img";
+import { EmailEventBox } from "./email-event-box";
 
 // To generate an example: make generate && go run ./backend mail > event-cancellation.html
 
@@ -22,12 +23,7 @@ export const EventCancellationEmail = () => (
       <Preview>Cancellation notice for {"{{.EventName}}"}</Preview>
       <Container style={container}>
         {/* Event Banner */}
-        <Img
-          alt="Event presentation"
-          src={"{{.ImageURL}}"}
-          width={960}
-          height={540}
-        />
+        <EmailEventImg src="{{.ImageURL}}" />
 
         {/* Black cancellation banner */}
         <Section style={welcome.section}>
@@ -47,16 +43,18 @@ export const EventCancellationEmail = () => (
             </Column>
           </Row>
           <Row>
-            <Column style={details.boxContainerLeft}>
-              <Box
+            <Column>
+              <EmailEventBox
                 title="PLANNED DATES"
                 icon={"{{.CalendarIconURL}}"}
                 iconAlt="Calendar icon"
                 content={`{{.EventStartDate}} → {{.EventEndDate}}`}
               />
             </Column>
-            <Column style={details.boxContainerRight}>
-              <Box
+          </Row>
+          <Row>
+            <Column>
+              <EmailEventBox
                 title="ADDRESS"
                 icon={"{{.PinIconURL}}"}
                 iconAlt="Pin icon"
@@ -77,40 +75,6 @@ export const EventCancellationEmail = () => (
   </Html>
 );
 
-const Box = (props: {
-  title: string;
-  icon: string;
-  iconAlt: string;
-  content: string;
-}) => {
-  return (
-    <Section style={box}>
-      <Row style={{ height: "100%" }}>
-        <Column style={{ height: "100%", verticalAlign: "top" }}>
-          <Section>
-            <Row>
-              <Column>
-                <Text style={boxTitle}>{props.title}</Text>
-              </Column>
-            </Row>
-            <Row>
-              <Column>
-                <Img
-                  alt={props.iconAlt}
-                  src={props.icon}
-                  width={32}
-                  height={32}
-                />
-                <Text style={boxContent}>{props.content}</Text>
-              </Column>
-            </Row>
-          </Section>
-        </Column>
-      </Row>
-    </Section>
-  );
-};
-
 export default EventCancellationEmail;
 
 // Styles
@@ -124,8 +88,7 @@ const main = {
 
 const container = {
   margin: "10px auto",
-  width: "600px",
-  maxWidth: "100%",
+  maxWidth: 800,
   border: "1px solid #F5F5F5",
 };
 
@@ -134,6 +97,7 @@ const welcome = {
     padding: "48px 20px",
     height: 220,
     backgroundColor: "#000000",
+    wordBreak: "break-word",
   },
   text: {
     color: "#FFFFFF",
@@ -144,38 +108,6 @@ const welcome = {
     lineHeight: 1.1,
     letterSpacing: -1.2,
   },
-} as const;
-
-const box = {
-  backgroundColor: "#F5F5F5",
-  borderRadius: 4,
-  padding: 12,
-  height: "100%",
-} as const;
-
-const boxTitle = {
-  margin: 0,
-  color: "#666666",
-  fontSize: 12,
-  fontWeight: 500,
-  letterSpacing: 0.5,
-  paddingBottom: 40,
-  lineHeight: 1.3,
-} as const;
-
-const boxContent = {
-  margin: 0,
-  fontSize: 16,
-  fontWeight: 500,
-  lineHeight: 1.3,
-  letterSpacing: -0.2,
-  paddingTop: 10,
-} as const;
-
-const boxContainerCommon = {
-  height: "100%",
-  width: "50%",
-  verticalAlign: "top",
 } as const;
 
 const details = {
@@ -197,14 +129,6 @@ const details = {
     letterSpacing: -0.6,
     margin: 0,
     marginBottom: 20,
-  },
-  boxContainerLeft: {
-    paddingRight: 8,
-    ...boxContainerCommon,
-  },
-  boxContainerRight: {
-    paddingLeft: 8,
-    ...boxContainerCommon,
   },
 } as const;
