+++ emails/tickets-confirmation.tmpl.tsx
@@ -6,13 +6,14 @@ import {
   Head,
   Heading,
   Html,
-  Img,
   Preview,
   Row,
   Section,
   Text,
 } from "@react-email/components";
 import React from "react";
+import { EmailEventImg } from "./email-event-img";
+import { EmailEventBox } from "./email-event-box";
 
 // To generate an example, run `make generate && go run ./backend mail > tickets-confirmation.html`
 
@@ -22,12 +23,7 @@ export const TicketsConfirmationEmail = () => (
     <Body style={main}>
       <Preview>Tickets for {"{{.EventName}}"}</Preview>
       <Container style={container}>
-        <Img
-          alt="Event presentation"
-          src="{{.ImageURL}}"
-          width={960}
-          height={540}
-        />
+        <EmailEventImg src="{{.ImageURL}}" />
         <Section style={welcome.section}>
           <Text style={welcome.text}>{"{{.WelcomeText}}"}</Text>
         </Section>
@@ -43,16 +39,18 @@ export const TicketsConfirmationEmail = () => (
             </Column>
           </Row>
           <Row>
-            <Column style={details.boxContainerLeft}>
-              <Box
+            <Column>
+              <EmailEventBox
                 title="DATE AND TIME"
                 icon={`{{.CalendarIconURL}}`}
                 iconAlt="Calendar icon"
                 content={`{{.TimeText}}`}
               />
             </Column>
-            <Column style={details.boxContainerRight}>
-              <Box
+          </Row>
+          <Row>
+            <Column>
+              <EmailEventBox
                 title="ADDRESS"
                 icon={`{{.PinIconURL}}`}
                 iconAlt="Pin icon"
@@ -73,40 +71,6 @@ export const TicketsConfirmationEmail = () => (
   </Html>
 );
 
-const Box = (props: {
-  title: string;
-  icon: string;
-  iconAlt: string;
-  content: string;
-}) => {
-  return (
-    <Section style={box}>
-      <Row style={{ height: "100%" }}>
-        <Column style={{ height: "100%", verticalAlign: "top" }}>
-          <Section>
-            <Row>
-              <Column>
-                <Text style={boxTitle}>{props.title}</Text>
-              </Column>
-            </Row>
-            <Row>
-              <Column>
-                <Img
-                  alt={props.iconAlt}
-                  src={props.icon}
-                  width={32}
-                  height={32}
-                />
-                <Text style={boxContent}>{props.content}</Text>
-              </Column>
-            </Row>
-          </Section>
-        </Column>
-      </Row>
-    </Section>
-  );
-};
-
 export default TicketsConfirmationEmail;
 
 const main = {
@@ -118,8 +82,7 @@ const main = {
 
 const container = {
   margin: "10px auto",
-  width: "600px",
-  maxWidth: "100%",
+  maxWidth: 800,
   border: "1px solid #F5F5F5",
 };
 
@@ -128,6 +91,7 @@ const welcome = {
     padding: "48px 20px",
     height: 220,
     backgroundColor: "#000000",
+    wordBreak: "break-word",
   },
   text: {
     color: "#FFFFFF",
@@ -138,38 +102,6 @@ const welcome = {
     lineHeight: 1.1,
     letterSpacing: -1.2,
   },
-} as const;
-
-const box = {
-  backgroundColor: "#F5F5F5",
-  borderRadius: 4,
-  padding: 12,
-  height: "100%",
-} as const;
-
-const boxTitle = {
-  margin: 0,
-  color: "#666666",
-  fontSize: 12,
-  fontWeight: 500,
-  letterSpacing: 0.5,
-  paddingBottom: 40,
-  lineHeight: 1.3,
-} as const;
-
-const boxContent = {
-  margin: 0,
-  fontSize: 16,
-  fontWeight: 500,
-  lineHeight: 1.3,
-  letterSpacing: -0.2,
-  paddingTop: 10,
-} as const;
-
-const boxContainerCommon = {
-  height: "100%",
-  width: "50%",
-  verticalAlign: "top",
 } as const;
 
 const details = {
@@ -191,14 +123,6 @@ const details = {
     letterSpacing: -0.6,
     margin: 0,
     marginBottom: 20,
-  },
-  boxContainerLeft: {
-    paddingRight: 8,
-    ...boxContainerCommon,
-  },
-  boxContainerRight: {
-    paddingLeft: 8,
-    ...boxContainerCommon,
   },
   seeEventButton: {
     backgroundColor: "#000000",
